Refactoring of core modules
