+++ app/src/main/java/pro/devapp/walkietalkiek/app/di/legacyAppDi.kt
@@ -3,17 +3,11 @@ package pro.devapp.walkietalkiek.app.di
 import org.koin.core.module.Module
 import org.koin.core.module.dsl.singleOf
 import pro.devapp.walkietalkiek.app.ChanelController
-import pro.devapp.walkietalkiek.app.Resolver
 import pro.devapp.walkietalkiek.app.SocketClient
 import pro.devapp.walkietalkiek.app.SocketServer
-import pro.devapp.walkietalkiek.serivce.network.data.ConnectedDevicesRepository
-import pro.devapp.walkietalkiek.serivce.network.data.DeviceInfoRepository
 
 fun Module.registerLegacyAppDi() {
-    singleOf(::ConnectedDevicesRepository)
-    singleOf(::DeviceInfoRepository)
     singleOf(::SocketClient)
     singleOf(::SocketServer)
     singleOf(::ChanelController)
-    singleOf(::Resolver)
 }
