+++ core-ui/build.gradle.kts
@@ -5,7 +5,7 @@ plugins {
 }
 
 android {
-    namespace = "pro.devapp.walkietalkiek.feature.ptt"
+    namespace = "pro.devapp.walkietalkiek.core"
     compileSdk = 35
 
     defaultConfig {
@@ -36,20 +36,12 @@ android {
 dependencies {
 
     implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.lifecycle.runtime.ktx)
-    implementation(platform(libs.androidx.compose.bom))
+    implementation(libs.androidx.viewmodel)
     implementation(libs.androidx.ui)
     implementation(libs.androidx.ui.graphics)
     implementation(libs.androidx.ui.tooling.preview)
     implementation(libs.androidx.material3)
-    implementation(libs.androidx.material3.adaptive.navigation.suite.android)
-    implementation(libs.koin.core)
-    implementation(libs.koin.android)
-    implementation(libs.koin.android.compose)
-
-    implementation(project(":core"))
-    implementation(project(":core-ui"))
-    implementation(project(":serivce-network"))
+    implementation(platform(libs.androidx.compose.bom))
 
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
