+++ serivce-network/build.gradle.kts
@@ -34,8 +34,13 @@ android {
 
 dependencies {
 
+    implementation(project(":core"))
+
     implementation(libs.androidx.core.ktx)
-    implementation(libs.androidx.appcompat)
+    implementation(libs.timber)
+    implementation(libs.koin.core)
+    implementation(libs.koin.android)
+
     implementation(libs.material)
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.junit)
